Adopt functional form for quark mass guesses
Due to the chiral restoration, for higher values of quark
chemical potentials, the determination of the quark masses
and renormalized chemical potentials may be quit problematic:
At low values, we must use a guess of around 300 to 350, but
at high values the chiral restoration takes place and we need
to use zero as a guess. To accomplish that, we use
     m = h / (1 + exp((mu - w) / t))
(h = height, w = width, t = transition width). Here mu is the
mean of the chemical potentials of the quarks.

This functional form may be more complex than necessary, but
at least it is quite smooth. The guesses are working better with
widths that exceed the values of chiral restoration; I believe that the
chiral restoration occurs at larger chemical values for asymmetric
matter.