Start transition to run on chemical potentials
Originally, the running parameter was the barionic density and
proton fraction. The idea is to transition to run on barionic and
isovector chemical potentials. I believe this will make possible
to properly determine transitions at lower values of pressure,
barionic chemical potential, and proton fractions.

The binodal will be a series of points in a barionic chemical
potential by isovector chemical potential plane.