Use grid method for calculation of binodals
There is a region in which the binodal makes a curve in the
barionic vs isovector chemical potentials space, making inpossible
to write the barionic chemical potential at transition as a
function of the isovector chemical potential. To make the
numerical calculation possible, we changed the approach to a grid
of barionic and isovector chemical potentials in which the
pressure is calculated. If the absolute value of the difference
in pressures at the grid point is less than a predetermined value,
we store that point as a binodal point.